chore: moving price logic to price contract.
