+++ contracts/packages/protocol/src/lib.rs
@@ -5,8 +5,9 @@ pub mod errors;
 pub mod guards;
 pub mod hub;
 pub mod hub_utils;
+pub use kujira;
 pub mod offer;
+pub mod price;
 pub mod profile;
 pub mod trade;
 pub mod trading_incentives;
-pub use kujira;
